requirements.txt
openai>=1.0.0
streamlit>=1.44.0
pandas>=2.2.0
matplotlib>=3.8.0
python-dotenv>=1.0.1
